Добавлены кнопки основного меню

+++ src/main/resources/liquibase/scripts/create_table__users_message_counter.sql
@@ -6,8 +6,8 @@ DROP TABLE IF EXISTS users_message_counter;
 
 CREATE TABLE user_message_counter
 (
-    id      INT NOT NULL,
-    user_id INT NOT NULL,
+    id      BIGINT NOT NULL,
+    user_id BIGINT NOT NULL,
     counter INT NOT NULL,
     PRIMARY KEY (id)
 );
